Reduce diff size
This will make the build script to alphabetical list the packages
on the `index.json`, which will reduce the diff size of the daily update.

This approach provides a nice side-effect that every day the diff
will only contain those packages that were added or removed since
the last update.